Split the api file into one file per entity type.
