+++ frontend/package.json
@@ -30,7 +30,7 @@
     "@types/bootstrap": "^5.2.10",
     "@types/jest": "^29.5.14",
     "@types/luxon": "^3.6.2",
-    "@types/react": "^19.1.0",
+    "@types/react": "^19.1.1",
     "@types/react-dom": "^19.1.2",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
